# src/main/java/com/senla/internship/classifiedapi/model/advertisement/Location.java
package com.senla.internship.classifiedapi.model.advertisement;

import com.fasterxml.jackson.annotation.JsonIgnore;
import lombok.*;

import javax.persistence.*;

/**
 * The {@code Location} class represents a location table.
 * @author Vlas Nagibin
 * @version 1.0
 */
@Getter
@Setter
@Entity
@ToString
@NoArgsConstructor
@AllArgsConstructor
@Table(name = "location")
public class Location {
    @Id
    @JsonIgnore
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(name = "city", nullable = false)
    private String city;

    @Column(name = "state", nullable = false)
    private String state;
}